Catch bus.sendCommand() errors and emit as warning events.
If we don't return the promise result of bus.sendCommand(), any
potential errors will be lost forever in the land of asychrony and
promises. A better way with "send and forget" actions which return a
promise, like bus.sendCommand(), is to `.catch()` any potential errors
and emit them for any listeners who might be interested.

 Changes to be committed:
	modified:   lib/services/catalog/commands/index.js
	modified:   lib/stores/redis-search/index.js
	modified:   lib/stores/reds/index.js